utils/requirements.txt
# Python/pip requirements for OpenRVDAS. Invoked on installation by the
# script utils/install_openrvdas.sh. See also requirements_mysql.txt for
# files required to enable MySQL functionality with OpenRVDAS

# Flags
--trusted-host pypi.org
--trusted-host files.pythonhosted.org

# Packages
pyserial
uwsgi
websockets
PyYAML
parse
psutil
setproctitle
influxdb_client
pyModbusTCP
GitPython

# For Geofencing
geopandas
shapely

django
djangorestframework

# Version-restricted packages
#django==5.0.3
#djangorestframework==3.15.1

#includes the swagger ui css and javascript
drf-spectacular[sidecar]

